+++ changelog.md
@@ -1,3 +1,14 @@
+# 1.1.0
+- add update-config
+    - keepAngularMayorVersion default true
+    - removeVersioningSymbols default false
+    - ignoreDependencies 
+    - ignoreDevDependencies
+- switch to typescript
+- remove postinstall (dose not work after npm 6) -> add a script. Not implement now
+- more unit tests, broken integration test
+- sonarqube fixes
+
 # no release
 - update integrationtests
 - fix pipeline